1.test each algorithm 2.organize the algorithms
